feat: 추천 검색어 가져오는 Custom Hook 생성
