+++ src/pages/App.tsx
@@ -1,7 +1,5 @@
-import { useEffect, useState } from 'react';
-import { getRecommendations } from '../api';
-import { Recommendation } from '../@types/recommendation';
-import debounce from '../utils/debounce';
+import { useState } from 'react';
+import useRecommendations from '../hooks/useRecommendations';
 import SearchButton from '../components/button';
 import SearchInput from '../components/input/SearchInput';
 import { Item, ListContainer, NoResultsItem } from '../components/list';
@@ -10,28 +8,7 @@ import * as S from './style';
 
 const App = () => {
   const [searchText, setSearchText] = useState('');
-  const [recommendations, setRecommendations] = useState<Recommendation[]>([]);
-
-  useEffect(() => {
-    if (searchText.trim() !== '') {
-      const controller = new AbortController();
-      const { signal } = controller;
-
-      const debounced = debounce(async () => {
-        const res = await getRecommendations(searchText, { signal });
-        if (res.isSuccess) {
-          setRecommendations(res.data);
-        }
-      }, 500);
-      debounced();
-
-      return () => {
-        debounced.clear();
-        controller.abort();
-      };
-    }
-    return undefined;
-  }, [searchText]);
+  const recommendations = useRecommendations({ name: searchText });
 
   return (
     <S.Wrapper>
